+++ session-2016-04-idf/coefs/tp3_exceptions_etape03.cpp
@@ -53,7 +53,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { throw Echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
