+++ session-2016-04-idf/coefs/tp4_generique_etape08.cpp
@@ -65,7 +65,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { throw EchecDivisionParZero() ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
